fix debug workflow

+++ scripts/Windows/2_deploy.sh
@@ -25,4 +25,6 @@ windeployqt $BUNDLE_PATH/lib/plugins/filter_sketchfab.dll --libdir $BUNDLE_PATH
 
 rm $BUNDLE_PATH/d3dcompiler*
 rm $BUNDLE_PATH/opengl32sw.dll
-rm $BUNDLE_PATH/vc_redist*
+if ls $BUNDLE_PATH/vc_redist* 1> /dev/null 2>&1; then
+    rm $BUNDLE_PATH/vc_redist*
+fi
